Fix bundle of issue for ICC build
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@6740 6f19259b-4bc3-4df7-8a09-765794883524

+++ DuetPkg/DxeIpl/Debug.c
@@ -57,9 +57,9 @@ PrintValue (
   UINT8  Char;
 
   for (Index = 0; Index < 8; Index++) {
-    Char = (UINT8)((Value >> ((7 - Index) * 4)) & 0x0f) + '0';
+    Char = (UINT8)(((Value >> ((7 - Index) * 4)) & 0x0f) + '0');
     if (Char > '9') {
-      Char = Char - '0' - 10 + 'A';
+      Char = (UINT8) (Char - '0' - 10 + 'A');
     }
     *mCursor = Char;
     mCursor += 2;
